pyproject.toml
[project]
name = "market_conditions"
version = "1.0.0"
description = "基于ccxt pro封装的实时行情数据收集和读取工具"
authors = [{name = "Market Conditions Team"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "ccxt>=4.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/market_conditions"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
